Implement compliments reliant on Google API and referring to a new sheet shared by Jason.
