+++ tests/unit/test_commands.py
@@ -26,14 +26,6 @@ def onetrue(*args):
     return len(truthiness) == 1
 
 
-@pytest.fixture
-def google_api_key(monkeypatch):
-    key = os.environ.get('GOOGLE_API_KEY')
-    if not key:
-        pytest.skip("Need GOOGLE_API_KEY environment variable")
-    monkeypatch.setitem(pmxbot.config, 'Google API key', key)
-
-
 class TestCommands:
     @classmethod
     def setup_class(cls):
